slim_ngd: Serialize shutdown sequence during SSR down
During SSR down, work is queued for shutdown
sequence. Sometimes work is getting scheduled
on one CPU while SSR up sequence is running
on other CPU. This is leading to unexpecpted
scenarios and ending with crash.
To overcome this, serialize shutdown sequence
by doing function call instead of workqueue.
And remove ctrl_up signal reset in the shutdown
sequence as it is affecting the transactions
initiated by client.

Change-Id: If73999ad19441e2090c4ad181dbc90e62ad3d9d4
Signed-off-by: Dilip Kota <dkota@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Laster K. (lazerl0rd) <officiallazerl0rd@gmail.com>