Fix all GCC 6.x warnings
Should be no changed functionality. The most problematic warnings were
the misleading indentation warning because they were often amongst
fairly dense blocks of code (particularly Qualcomm drivers), making
figuring out if the mis-indented block was supposed to be wrapped in
the if/for statement or not. I decided to play it safe and just reset
the indentation as it will keep functionality with how it is before the
cleanup (which wasn't obviously broken). If you think there is a block
that is wrong, feel free to comment and let me know.

The array-bounds warning in net/ipv4/tcp_input.c is a false warning,
there is no way for an array out of bounds to be triggered. This
block is the same in upstream at the time of authoring this.

Warnings handled:
-Warray-bounds
-Wbool-compare
-Wmemset-transposed-args
-Wmisleading-indentation
-Wtautological-compare

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Laster K. (lazerl0rd) <officiallazerl0rd@gmail.com>